understand the react life cycle component
1. for the componentDidMount()
once the state is updated, the render() method would call the render method again
2. componentWillUnmount()
it is called before a component is unmounted, a component is removed from the view. if I have allocated for the component resources or things unallocated them in this specific method
3. shouldComponentUpdate(nextprops,nextState)
the return value is boolen ,
if it is the false that it returns, even if the state is updated , the render will not be called .
if it is the true that is returns the view would be updated.
it is up to you whether the view would update if the state is updated